Generate and render pages for available locales
Each generated page is aware of its parent document (canon) and its
sibling locale page, if the site supports more than two locales, via the
`hreflangs` attribute.

The `content`, `data` and other attributes of the generated pages mirror
that of the associated canon page or document. The difference lies in the
`url` attribute, which is simply the url of canon object prepended by
current locale. The Liquid context for rendering is also enhanced to point
to current locale and the prepended url.